consolidate key.go -> keys.go, core change to getValidatorPowerRank
